+++ src/main/kotlin/dartzee/ai/AbstractDartsSimulation.kt
@@ -13,7 +13,7 @@ import java.sql.Timestamp
 
 abstract class AbstractDartsSimulation(protected var dartboard: Dartboard,
                                        protected var player: PlayerEntity,
-                                       protected var model: DartsAiModelMk2) : DartboardListener
+                                       protected var model: DartsAiModel) : DartboardListener
 {
     //Transient things
     protected var dtStart: Timestamp? = null
