+++ src/main/kotlin/dartzee/screen/ai/AIConfigurationSubPanelDartzee.kt
@@ -1,6 +1,6 @@
 package dartzee.screen.ai
 
-import dartzee.ai.DartsAiModelMk2
+import dartzee.ai.DartsAiModel
 import dartzee.ai.DartzeePlayStyle
 import dartzee.core.bean.RadioButtonPanel
 import javax.swing.JRadioButton
@@ -20,13 +20,13 @@ class AIConfigurationSubPanelDartzee : AbstractAIConfigurationSubPanel()
 
     override fun valid() = true
 
-    override fun populateModel(model: DartsAiModelMk2): DartsAiModelMk2
+    override fun populateModel(model: DartsAiModel): DartsAiModel
     {
         val dartzeePlayStyle = if (rdbtnCautious.isSelected) DartzeePlayStyle.CAUTIOUS else DartzeePlayStyle.AGGRESSIVE
         return model.copy(dartzeePlayStyle = dartzeePlayStyle)
     }
 
-    override fun initialiseFromModel(model: DartsAiModelMk2)
+    override fun initialiseFromModel(model: DartsAiModel)
     {
         rdbtnCautious.isSelected = model.dartzeePlayStyle == DartzeePlayStyle.CAUTIOUS
         rdbtnAggressive.isSelected = model.dartzeePlayStyle == DartzeePlayStyle.AGGRESSIVE
