+++ src/main/kotlin/dartzee/screen/ai/AbstractAIConfigurationSubPanel.kt
@@ -1,16 +1,16 @@
 package dartzee.screen.ai
 
-import dartzee.ai.DartsAiModelMk2
+import dartzee.ai.DartsAiModel
 import javax.swing.JPanel
 
 abstract class AbstractAIConfigurationSubPanel : JPanel()
 {
     abstract fun valid(): Boolean
-    abstract fun populateModel(model: DartsAiModelMk2): DartsAiModelMk2
-    abstract fun initialiseFromModel(model: DartsAiModelMk2)
+    abstract fun populateModel(model: DartsAiModel): DartsAiModel
+    abstract fun initialiseFromModel(model: DartsAiModel)
 
     fun reset()
     {
-        initialiseFromModel(DartsAiModelMk2.new())
+        initialiseFromModel(DartsAiModel.new())
     }
 }
