+++ src/main/kotlin/dartzee/screen/ai/VisualisationPanelDensity.kt
@@ -1,7 +1,7 @@
 package dartzee.screen.ai
 
 import dartzee.`object`.ColourWrapper
-import dartzee.ai.DartsAiModelMk2
+import dartzee.ai.DartsAiModel
 import dartzee.core.bean.paint
 import dartzee.utils.DartsColour
 import dartzee.utils.getDistance
@@ -30,7 +30,7 @@ class VisualisationPanelDensity: AbstractVisualisationPanel()
         panelKey.icon = ImageIcon(keyImg)
     }
 
-    override fun showVisualisation(hmPointToCount: Map<Point, Int>, model: DartsAiModelMk2)
+    override fun showVisualisation(hmPointToCount: Map<Point, Int>, model: DartsAiModel)
     {
         val centerPt = model.getScoringPoint(dartboard)
