+++ src/main/kotlin/dartzee/screen/game/dartzee/GamePanelDartzee.kt
@@ -2,7 +2,6 @@ package dartzee.screen.game.dartzee
 
 import dartzee.`object`.Dart
 import dartzee.ai.DartsAiModel
-import dartzee.ai.DartsAiModelMk2
 import dartzee.core.obj.HashMapList
 import dartzee.dartzee.DartzeeRoundResult
 import dartzee.dartzee.DartzeeRuleDto
@@ -39,7 +38,7 @@ class GamePanelDartzee(parent: AbstractDartsGameScreen,
     override fun factoryDartboard() = DartzeeDartboard()
     override fun factoryState(pt: ParticipantEntity, scorer: DartsScorerDartzee) = DartzeePlayerState(pt, scorer)
 
-    override fun doAiTurn(model: DartsAiModelMk2)
+    override fun doAiTurn(model: DartsAiModel)
     {
         if (isScoringRound())
         {
